# uni_core/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: 企业级 RBAC 权限管理系统后端 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RBAC Admin API
  version: 1.0.0
paths:
...
  /docker/hosts/{id}/cmds/{ref}/pull:
    get:
      description: 按指令号接入拉取进度会话；每行一个进度记录 {"seq","t","id","status","current","total","done","error","eof"}；客户端断开即下发
        cancel（终止拉取）
      parameters:
      - description: 设备ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 指令号
        in: path
        name: ref
        required: true
        type: string
      produces:
      - application/x-ndjson
      responses:
        "200":
          description: 流已建立(逐行 NDJSON)
        "400":
          description: 请求参数不合法 / 该指令不支持拉取进度流
        "401":
          description: 未登录
        "403":
          description: 无权限 / 非发起人
        "404":
          description: 指令不存在或已过期
        "409":
          description: 该指令没有可接入的流会话 / 会话已有连接
      security:
      - BearerAuth: []
      summary: 镜像拉取进度流(NDJSON)
      tags:
      - Docker 管理
  /docker/hosts/{id}/cmds/{ref}/stats:
    get:
      description: 按指令号接入已建立的 stats 会话；每行一个样本 {"seq","t","cpu_percent","mem_usage_mb","mem_limit_mb","net_rx_bytes_sec","net_tx_bytes_sec","eof"}；客户端断开即下发
        cancel
      parameters:
      - description: 设备ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 指令号
        in: path
        name: ref
        required: true
        type: string
      produces:
      - application/x-ndjson
      responses:
        "200":
          description: 流已建立(逐行 NDJSON)
        "400":
          description: 请求参数不合法 / 该指令不支持 stats 流
        "401":
          description: 未登录
        "403":
          description: 无权限 / 非发起人
        "404":
          description: 指令不存在或已过期
        "409":
          description: 该指令没有可接入的流会话 / 会话已有连接
      security:
      - BearerAuth: []
      summary: 容器 stats 实时流(NDJSON)
      tags:
      - Docker 管理
  /docker/hosts/{id}/cmds/{ref}/stream:
    get:
      description: 按指令号接入已建立的日志流会话（container:logs follow 或 compose:logs 项目聚合日志）；每行
        {"seq","data"(base64),"eof"}；客户端断开即下发 cancel
      parameters:
      - description: 设备ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 指令号
        in: path
        name: ref
        required: true
        type: string
      produces:
      - application/x-ndjson
      responses:
        "200":
          description: 流已建立(逐行 NDJSON)
        "400":
          description: 请求参数不合法 / 该指令不支持日志流
        "401":
          description: 未登录
        "403":
          description: 无权限 / 非发起人
        "404":
          description: 指令不存在或已过期
        "409":
          description: 该指令没有可接入的流会话 / 会话已有连接
      security:
      - BearerAuth: []
      summary: 容器/项目聚合日志流(NDJSON Follow)
      tags:
      - Docker 管理
...
  /docker/hosts/{id}/stream/exec:
    get:
      description: 用 streamTicket 升级；服务端发 {"type":"data","data":base64} / {"type":"eof"}，客户端发
        input/resize/cancel
      parameters:
      - description: 设备ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 一次性票据(来自 cmds/:ref 结果)
        in: query
        name: ticket
        required: true
        type: string
      responses:
        "101":
          description: 握手成功，协议切换到 WebSocket
        "400":
          description: 参数不合法
        "401":
          description: 票据无效或已过期
        "403":
          description: 票据与主机不符
        "409":
          description: 该流会话已有连接
      summary: 容器终端(WebSocket)
      tags:
      - Docker 管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"